Fix crash in mcts-cpp
GLAS was passed in as a pointer, which can cause memory corruption if
Python decides to free the associated memory. Instead, we rely on a
copy of the GLAS objects now (otherwise we would need to add logic to
manually free the memory). This should not add much overhead, but
requires to always pass in a dummy GLAS object, even if GLAS is not
used.